+++ test/selenium/contributions/new_flow/pages/ContributionsLanding.scala
@@ -2,9 +2,9 @@ package selenium.contributions.new_flow.pages
 
 import org.openqa.selenium.WebDriver
 import org.scalatest.selenium.Page
-import selenium.util.{Browser, Config}
+import selenium.util.{Browser, Config, TestUser}
 
-case class ContributionsLanding(region: String)(implicit val webDriver: WebDriver) extends Page with Browser {
+case class ContributionsLanding(region: String, testUser: TestUser)(implicit val webDriver: WebDriver) extends Page with Browser {
 
   // While the new contributions landing page test is running,
   // we just want Selenium to test the old page.
@@ -19,6 +19,39 @@ case class ContributionsLanding(region: String)(implicit val webDriver: WebDrive
   private val otherAmountButton = cssSelector(".form__radio-group-label[for='contributionAmount-other']")
 
   private val otherAmount = id("contributionOther")
+
+  private val stripeSelector = cssSelector(".form__radio-group-label[for='paymentMethodSelector-Stripe']")
+  private val payPalSelector = cssSelector(".form__radio-group-label[for='paymentMethodSelector-PayPal']")
+  private val stateSelector = id("contributionState")
+
+  private object RegisterFields {
+    private val firstName = id("contributionFirstName")
+    private val lastName = id("contributionLastName")
+    private val email = id("contributionEmail")
+
+    def fillIn() {
+
+      setValue(email, s"${testUser.username}@gu.com", clear = true)
+      setValue(firstName, testUser.username, clear = true)
+      setValue(lastName, testUser.username, clear = true)
+    }
+
+    def clear(): Unit = {
+      clearValue(email)
+      clearValue(firstName)
+      clearValue(lastName)
+    }
+  }
+
+  def fillInPersonalDetails() { RegisterFields.fillIn() }
+
+  def clearForm(): Unit = RegisterFields.clear()
+
+  def selectState: Unit = setSingleSelectionValue(stateSelector, "NY")
+
+  def selectStripePayment(): Unit = clickOn(stripeSelector)
+
+  def selectPayPalPayment(): Unit = clickOn(payPalSelector)
 
   def pageHasLoaded: Boolean = {
     pageHasElement(contributeButton)
